move: shared 폴더로 prize card 이동
