+++ packages/user/src/components/shared/PrizeCard.tsx
